add buffer_csv (offline) mode, refactored the whole code, modified gitignore
