+++ src/evl/core/buffer_csv.hpp
@@ -1,0 +1,9 @@
+// Copyright 2018 Event Vision Library.
+#ifndef SRC_EVL_CORE_BUFFER_CSV_HPP_
+#define SRC_EVL_CORE_BUFFER_CSV_HPP_
+
+#include "types.hpp"
+
+void bufferData(EventBuffer *buffer, char* fname);
+
+#endif  // SRC_EVL_CORE_BUFFER_CSV_HPP_
