+++ src/evl/core/common.hpp
@@ -1,0 +1,9 @@
+// Copyright 2018 Event Vision Library.
+#ifndef SRC_EVL_CORE_COMMON_HPP_
+#define SRC_EVL_CORE_COMMON_HPP_
+
+#include <mutex>
+
+extern std::mutex mtx;
+
+#endif   // SRC_EVL_CORE_COMMON_HPP_
